+++ packages/wepy/test/fake/com_a.js
@@ -4,7 +4,15 @@ let wepy = require('../../lib/wepy.js').default;
 
 let ComAA = require('./com_a_a');
 
+
+function Car() {
+
+}
+Car.prototype.run = function () {}
+
 module.exports = class Com extends wepy.component {
+
+
     constructor () {
         super();
         this.data = {
@@ -21,6 +29,38 @@ module.exports = class Com extends wepy.component {
         this.components = {
             comaa: ComAA
         }
+
+        this.props = {
+            comprop1: String,
+            comprop2: [Number, String, Boolean],
+            comprop3: {
+                coerce: function (v) {
+                    return +v;
+                },
+                default: 50
+            },
+            comprop4: Object,
+            comprop5: {
+                default: 60
+            },
+            comprop6: {
+                type: [Number],
+                default: 70
+            },
+            comprop10: {
+                type: [Boolean, Function, Object, Array, Car],
+                coerce: function (v) {
+                    return +v;
+                },
+                default: function () {
+                    return 50
+                }
+            },
+        };
+
+        this.$props = {
+            comaa: { 'v-bind:comaaprop': 'comprop6' }
+        };
     }
 
     customMethod () {