+++ packages/wepy/test/fake/page.js
@@ -3,6 +3,7 @@ let assert = require('assert');
 let wepy = require('../../lib/wepy.js').default;
 let Mix = require('./mixin');
 let ComA = require('./com_a');
+let ComB = require('./com_b');
 
 module.exports = class Index extends wepy.page {
     constructor () {
@@ -13,12 +14,28 @@ module.exports = class Index extends wepy.page {
             }
         };
         this.components = {
-            coma: ComA
+            coma: ComA,
+            comb: ComB
         };
         this.data = {
-            'a': 1
+            'a': 1,
+            'objProp': {a:1},
+            'c': 'string',
+            'stringNumber': '123',
+            'testValid': '50'
         };
         this.mixins = [Mix];
+
+        this.$props = {
+            coma: {
+                'v-bind:comprop1': 'c',
+                'comprop2': 'static props',
+                'v-bind:comprop4': 'objProp',
+                'v-bind:comprop6': 'testValid',
+                'v-bind:comprop3': 'stringNumber',
+                'v-bind:comprop10': 'a'
+            }
+        };
     }
     onShow (args) {
         assert.strictEqual(this instanceof Index, true, 'page onShow triggered, this is self');